StatesTrajectory: changed from std::vector<> to Array_<>
This is for compatibility with Simbody.  SimTK::Array_<> is essentially the same (including most all method calls) as std::vector<> but with some performance and binary compatibility improvements.

The only changes that are needed in the rest of OpenSim is that some arguments are now unsigned int instead of size_t.