Adjustments to the StatesDocument API
The following was done to better fit with OpenSim::StatesTrajectory:

- Eliminated the default constructor StatesDocument().
- Added the following constructors: 1) StatesDocument(file), 2) StatesDocument(xmlDoc), and 3) StatesDocument(model, trajectory).
- serializeToFile(file) now only requires a file name.
- serializeToString(xmlDoc) now only requires a document string.
- There is now only a single deserialization method:  deserialize(model, trajectory)

All of these changes rely on the design strategy that a constructed StatesDocument always contains a complete and valid underlying XML Document.  Additionally, a StatesDocument never holds a Model or trajectory of states (i.e., Array_<State>) in a member variable.  Model and trajectory must always be supplied externally when needed.
-