Element::setValueAs(): removed precision argument
Previously, I added a default argument that allowed the precision of XML output (number of digits written) to be set.

Xml::Element::setValueAs<T>(T& time, int precision=LosslessNumDigitsReal)

This capability was abandoned so as not to introduce changes in the Simbody API. Should a need arise that would benefit from the new API, the precision argument could be taken up again.