Class name changed to ExponentialContactForce
The name of class ExponentialContact was changed to ExponentialContactForce. The motivation for doing this was mainly to reflect the fact the this class is hierarchically a Force, as opposed to just contact geometry for example.

The files names of the class files were changes as well:
ExponentialContact.h --> ExponentialContactForce.h
ExponentialContact.cpp --> ExponentialContactForce.cpp

The file names of the testing executables (testExponentialContact.cpp and testContactPerformance.cpp) were not changed.

The change of the class name had minor ripple effects (e.g., in type registration).

In addition, method separators (//______________) were deleted in favor of just using an empty line.

Finally, comments in .cpp files were trimmed when possible.

All test cases are passing.