Draft addition of copy constructor, copy assignment operator, and move constructor.
As of this commit, only the copy constructor has been tested, but all the original test cases are passing with the copy constructor in place.

The copy constructor was a bear to get working. The order in which things are done has to be just right!

The next step is to write a dedicated test case for the new constructors and the copy assignment operator.