Got rid of ExponentialSpringForce pointer in ExponentialContactForce.
Instead of holding on to the pointer to the underlying ExponentialSpringForce subystem (member variable was `_spr`), the code now gets a reference to the adopted subsystem.

I added two private methods to assist with this:
const ExponentialSpringForce& getSprRef() const, and
ExponentialSpringForce& updSprRef()

These methods simply perform a cast.

These changes also allowed the underlying ExponentialSpringForce instance to be created as a local variable, which means there doesn't need to be any cleanup of the previously dynamically allocated instance (i.e., no `delete _spr;` is needed).

The test cases appear to be running without memory leaks, except for one very small one, which I will track down.