Added asserts to check for valid Force indexes
The asserts are in the private helper methods getSprRef() and updSprRef(), which are used downcast the abstract Force Subsystem into an ExponentialSpringForce reference.

These methods should not be called prior to proper indices have been generated, which can happen if an ExponentialContactForce has been constructed but the underlying ExponentialSpringForce has not yet been instantiated. See the method `extendAddToSystem()`.