Got rid of the ExponentialSpringForce pointer.
The underlying ExponentialSpringForce is created as a local variable, instead of being allocated dynamically.
To access its SimTK::Subystem, its ForceIndex is used to get a reference.

Helper methods (getSprRef() and updSprRef()) were added to assist with acquiring the reference to the underlying ExponentialSpringForce Subsystem.

The explicit copy and move constructors have been commented out, along with the copy assignment operator, as the motivation for these methods was to handle the pointer properly.  Now that there is not pointer, the compiler-generated versions of these methods may work fine.