Store station as a subcomponent of ExponentialContactForce
This change mimics the implementation of the Joint class, where we now create and store the contact Station as an internal property, effectively making it a subcomponent. We then connect theSocket between the force and the Station enabling the existing downstream behavior.

Additional changes have been made including using static_cast to access the internal SimTK::ExponentialSpringForce and other various changes to make testExponentialContact pass.