Updated and polished general comments for ExponentialContactForce
There were a few minor changes. The main change was updating the example code for constructing an ExponentialContactForce instance. Previously, construction was done with a Station. Now, it's done with a PhysicalFrame and body location.

Still need to 1) proof the documentation comments for the methods and 2) possibly add a typical snippet from an .osim Model file.